requirements.txt
botbuilder-azure==4.14.6
botbuilder-core==4.14.6
botbuilder-schema==4.14.6